SGW-398 OnDemandPoll may stop dnp stack from executing tasks
Root cause:

SGW sends OnDemandPoll requests to dnp stack even when the stack is
closed. The master in dnp stack will generate a exception instead of
executing the command. But the master scheduler thinks the masters
is still executing the task so it won't schedule other tasks in the
queue to run.

Solution:

Don't run the OnDemandPoll Request if the dnp stack is closed. Simply
disable the task.

Verification:

The unit test is in progress in my testbed